refactor(local-community): extract challenge helpers to local-community/challenges.ts
Moves handleChallengeRequest, handleChallengeAnswer, handleChallengeExchange,
and all supporting helpers (_publishChallenges, _publishChallengeVerification,
_publishFailedChallengeVerification, _publishIdempotentDuplicateVerification,
_decryptOrRespondWithFailure, _parseChallengeRequestPublicationOrRespondWithFailure,
_parseChallengeAnswerOrRespondWithFailure, _buildRuntimeChallengeRequestPublication,
_buildRuntimeChallengeRequest, _cleanUpChallengeAnswerPromise,
_storePublicationAndEncryptForChallengeVerification) into a focused module.

Splits handleChallengeRequest into parseAndValidateChallengeRequest,
runChallengeExchangeIfNeeded, and runVerificationAndStorePublication with
handleChallengeRequest as a short orchestrator, without changing behavior.

Keeps handleChallengeRequest, handleChallengeAnswer, and handleChallengeExchange
as thin method facades on the class so the constructor bind on
handleChallengeExchange still works and the pubsub subscribe call remains
unchanged. Drops 'private' from _challengeAnswerPromises,
_challengeAnswerResolveReject, _ongoingChallengeExchanges,
_duplicatePublicationAttempts, and _challengeExchangesFromLocalPublishers so
the new free functions can access them. Updates publication-validation.ts to
call publishFailedChallengeVerification directly.
