+++ src/runtime/node/community/local-community.ts
@@ -246,6 +246,11 @@ import {
     isPublicationAuthorPartOfRoles,
     respondWithErrorIfSignatureOfPublicationIsInvalid
 } from "./local-community/publication-validation.js";
+import {
+    handleChallengeAnswer as handleChallengeAnswerFreeFunction,
+    handleChallengeExchange as handleChallengeExchangeFreeFunction,
+    handleChallengeRequest as handleChallengeRequestFreeFunction
+} from "./local-community/challenges.js";
 
 // This is a sub we have locally in our pkc datapath, in a NodeJS environment
 export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalCommunityParsedOptions {
@@ -257,14 +262,14 @@ export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalC
 
     // These caches below will be used to facilitate challenges exchange with authors, they will expire after 10 minutes
     // Most of the time they will be delete and cleaned up automatically
-    private _challengeAnswerPromises!: LRUCache<string, Promise<DecryptedChallengeAnswer["challengeAnswers"]>>;
-    private _challengeAnswerResolveReject!: LRUCache<
+    _challengeAnswerPromises!: LRUCache<string, Promise<DecryptedChallengeAnswer["challengeAnswers"]>>;
+    _challengeAnswerResolveReject!: LRUCache<
         string,
         { resolve: (answers: DecryptedChallengeAnswer["challengeAnswers"]) => void; reject: (error: Error) => void }
     >;
-    private _ongoingChallengeExchanges!: LRUCache<string, boolean>;
-    private _duplicatePublicationAttempts!: LRUCache<string, number>;
-    private _challengeExchangesFromLocalPublishers: Record<string, boolean> = {}; // key is stringified challengeRequestId and value is true if the challenge exchange is ongoing
+    _ongoingChallengeExchanges!: LRUCache<string, boolean>;
+    _duplicatePublicationAttempts!: LRUCache<string, number>;
+    _challengeExchangesFromLocalPublishers: Record<string, boolean> = {}; // key is stringified challengeRequestId and value is true if the challenge exchange is ongoing
 
     _cidsToUnPin: Set<string> = new Set<string>();
     _mfsPathsToRemove: Set<string> = new Set<string>();
@@ -803,640 +808,16 @@ export class LocalCommunity extends RpcLocalCommunity implements CreateNewLocalC
         }
     }
 
-    private async _decryptOrRespondWithFailure(request: ChallengeRequestMessageType | ChallengeAnswerMessageType): Promise<string> {
-        const log = Logger("pkc-js:local-community:_decryptOrRespondWithFailure");
-        try {
-            return await decryptEd25519AesGcmPublicKeyBuffer(request.encrypted, this.signer.privateKey, request.signature.publicKey);
-        } catch (e) {
-            log.error(`Failed to decrypt request (${request.challengeRequestId.toString()}) due to error`, e);
-            await this._publishFailedChallengeVerification(
-                { reason: messages.ERR_COMMUNITY_FAILED_TO_DECRYPT_PUBSUB_MSG },
-                request.challengeRequestId
-            );
-
-            throw e;
-        }
-    }
-
-    private async _publishChallenges(
-        challenges: Omit<Challenge, "verify">[],
-        request: DecryptedChallengeRequestMessageTypeWithCommunityAuthor
-    ) {
-        const log = Logger("pkc-js:local-community:_publishChallenges");
-        const toEncryptChallenge = <DecryptedChallenge>{ challenges };
-        const toSignChallenge: Omit<ChallengeMessageType, "signature"> = cleanUpBeforePublishing({
-            type: "CHALLENGE",
-            protocolVersion: env.PROTOCOL_VERSION,
-            userAgent: this._pkc.userAgent,
-            challengeRequestId: request.challengeRequestId,
-            encrypted: await encryptEd25519AesGcmPublicKeyBuffer(
-                deterministicStringify(toEncryptChallenge),
-                this.signer.privateKey,
-                request.signature.publicKey
-            ),
-            timestamp: timestamp()
-        });
-
-        const challengeMessage = <ChallengeMessageType>{
-            ...toSignChallenge,
-            signature: await signChallengeMessage({ challengeMessage: toSignChallenge, signer: this.signer })
-        };
-        const pubsubClient = this._clientsManager.getDefaultKuboPubsubClient();
-
-        this._clientsManager.updateKuboRpcPubsubState("publishing-challenge", pubsubClient.url);
-
-        // we only publish over pubsub if the challenge exchange is not ongoing for local publishers
-        if (!this._challengeExchangesFromLocalPublishers[request.challengeRequestId.toString()])
-            await this._clientsManager.pubsubPublish(pubsubTopicWithfallback(this), challengeMessage);
-        log(
-            `Community ${this.address} with pubsub topic ${pubsubTopicWithfallback(this)} published ${challengeMessage.type} over pubsub: `,
-            remeda.pick(toSignChallenge, ["timestamp"]),
-            toEncryptChallenge.challenges.map((challenge) => challenge.type)
-        );
-        this._clientsManager.updateKuboRpcPubsubState("waiting-challenge-answers", pubsubClient.url);
-        this.emit("challenge", {
-            ...challengeMessage,
-            challenges
-        });
-    }
-
-    async _publishFailedChallengeVerification(
-        result: Pick<ChallengeVerificationMessageType, "challengeErrors" | "reason">,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"]
-    ) {
-        // challengeSucess=false
-        const log = Logger("pkc-js:local-community:_publishFailedChallengeVerification");
-
-        const toSignVerification: Omit<ChallengeVerificationMessageType, "signature"> = cleanUpBeforePublishing({
-            type: "CHALLENGEVERIFICATION",
-            challengeRequestId: challengeRequestId,
-            challengeSuccess: false,
-            challengeErrors: result.challengeErrors,
-            reason: result.reason,
-            userAgent: this._pkc.userAgent,
-            protocolVersion: env.PROTOCOL_VERSION,
-            timestamp: timestamp()
-        });
-
-        const challengeVerification = <ChallengeVerificationMessageType>{
-            ...toSignVerification,
-            signature: await signChallengeVerification({ challengeVerification: toSignVerification, signer: this.signer })
-        };
-
-        const pubsubClient = this._clientsManager.getDefaultKuboPubsubClient();
-        this._clientsManager.updateKuboRpcPubsubState("publishing-challenge-verification", pubsubClient.url);
-        log(
-            `Will publish ${challengeVerification.type} over pubsub topic ${pubsubTopicWithfallback(this)} on community ${this.address}:`,
-            remeda.omit(toSignVerification, ["challengeRequestId"])
-        );
-
-        if (!this._challengeExchangesFromLocalPublishers[challengeRequestId.toString()])
-            await this._clientsManager.pubsubPublish(pubsubTopicWithfallback(this), challengeVerification);
-        this._clientsManager.updateKuboRpcPubsubState("waiting-challenge-requests", pubsubClient.url);
-
-        this.emit("challengeverification", challengeVerification);
-        this._ongoingChallengeExchanges.delete(challengeRequestId.toString());
-        delete this._challengeExchangesFromLocalPublishers[challengeRequestId.toString()];
-        this._cleanUpChallengeAnswerPromise(challengeRequestId.toString());
-    }
-
-    private async _publishIdempotentDuplicateVerification(
-        request: DecryptedChallengeRequestMessageType,
-        challengeRequestId: ChallengeRequestMessageType["challengeRequestId"],
-        duplicateReason: string
-    ) {
-        const log = Logger("pkc-js:local-community:_publishIdempotentDuplicateVerification");
-
-        let encrypted: ChallengeVerificationMessageType["encrypted"] | undefined;
-        let toEncryptDecrypted: DecryptedChallengeVerification | undefined;
-
-        // For comments, include the existing comment data in the encrypted response
-        if (duplicateReason === messages.ERR_DUPLICATE_COMMENT && request.comment) {
-            const existingComment = this._dbHandler.queryCommentBySignatureEncoded(request.comment.signature.signature);
-            if (!existingComment) {
-                return this._publishFailedChallengeVerification({ reason: duplicateReason }, challengeRequestId);
-            }
-            log("Returning idempotent success for duplicate comment", existingComment.cid);
-
-            const authorSignerAddress = await getPKCAddressFromPublicKey(existingComment.signature.publicKey);
-            const authorDomain = getAuthorNameFromWire(existingComment.author);
-            const authorCommunity = this._dbHandler.queryCommunityAuthor(authorSignerAddress, authorDomain);
-            if (!authorCommunity) {
-                return this._publishFailedChallengeVerification({ reason: duplicateReason }, challengeRequestId);
-            }
-            const commentNumberPostNumber = this._dbHandler._assignNumbersForComment(existingComment.cid);
-
-            const commentUpdateNoSig = <Omit<DecryptedChallengeVerification["commentUpdate"], "signature">>cleanUpBeforePublishing({
-                author: { community: authorCommunity },
-                cid: existingComment.cid,
-                protocolVersion: env.PROTOCOL_VERSION,
-                ...commentNumberPostNumber
-            });
-            const commentUpdate = <DecryptedChallengeVerification["commentUpdate"]>{
-                ...commentUpdateNoSig,
-                signature: await signCommentUpdateForChallengeVerification({
-                    update: commentUpdateNoSig,
-                    signer: this.signer
-                })
-            };
-            const commentIpfs = CommentIpfsSchema.strip().parse(existingComment);
-            toEncryptDecrypted = { comment: commentIpfs, commentUpdate };
-
-            encrypted = await encryptEd25519AesGcmPublicKeyBuffer(
-                deterministicStringify(toEncryptDecrypted),
-                this.signer.privateKey,
-                request.signature.publicKey
-            );
-        } else {
-            // For edits/moderations: success has no encrypted data (same as normal success)
-            log("Returning idempotent success for duplicate", duplicateReason);
-        }
-
-        const toSignMsg: Omit<ChallengeVerificationMessageType, "signature"> = cleanUpBeforePublishing({
-            type: "CHALLENGEVERIFICATION",
-            challengeRequestId,
-            encrypted,
-            challengeSuccess: true,
-            reason: undefined,
-            userAgent: this._pkc.userAgent,
-            protocolVersion: env.PROTOCOL_VERSION,
-            timestamp: timestamp()
-        });
-        const challengeVerification = <ChallengeVerificationMessageType>{
-            ...toSignMsg,
-            signature: await signChallengeVerification({ challengeVerification: toSignMsg, signer: this.signer })
-        };
-
-        const pubsubClient = this._clientsManager.getDefaultKuboPubsubClient();
-        this._clientsManager.updateKuboRpcPubsubState("publishing-challenge-verification", pubsubClient.url);
-        if (!this._challengeExchangesFromLocalPublishers[challengeRequestId.toString()])
-            await this._clientsManager.pubsubPublish(pubsubTopicWithfallback(this), challengeVerification);
-        this._clientsManager.updateKuboRpcPubsubState("waiting-challenge-requests", pubsubClient.url);
-
-        const objectToEmit = <DecryptedChallengeVerificationMessageType>{ ...challengeVerification, ...toEncryptDecrypted };
-        this.emit("challengeverification", objectToEmit);
-        this._ongoingChallengeExchanges.delete(challengeRequestId.toString());
-        delete this._challengeExchangesFromLocalPublishers[challengeRequestId.toString()];
-        this._cleanUpChallengeAnswerPromise(challengeRequestId.toString());
-    }
-
-    private async _storePublicationAndEncryptForChallengeVerification(
-        request: DecryptedChallengeRequestMessageType,
-        pendingApproval?: boolean
-    ): Promise<(DecryptedChallengeVerification & Required<Pick<DecryptedChallengeVerificationMessageType, "encrypted">>) | undefined> {
-        const commentAfterAddingToIpfs = await storePublication(this, request, pendingApproval);
-        if (!commentAfterAddingToIpfs) return undefined;
-        const authorSignerAddress = await getPKCAddressFromPublicKey(commentAfterAddingToIpfs.comment.signature.publicKey);
-        const authorDomain = getAuthorNameFromWire(commentAfterAddingToIpfs.comment.author);
-
-        const authorCommunity = this._dbHandler.queryCommunityAuthor(authorSignerAddress, authorDomain);
-        if (!authorCommunity) throw Error("author.community can never be undefined after adding a comment");
-        const commentNumberPostNumber = this._dbHandler._assignNumbersForComment(commentAfterAddingToIpfs.cid);
-
-        const commentUpdateOfVerificationNoSignature = <Omit<DecryptedChallengeVerification["commentUpdate"], "signature">>(
-            cleanUpBeforePublishing({
-                author: { community: authorCommunity },
-                cid: commentAfterAddingToIpfs.cid,
-                protocolVersion: env.PROTOCOL_VERSION,
-                pendingApproval,
-                ...commentNumberPostNumber
-            })
-        );
-        const commentUpdate = <DecryptedChallengeVerification["commentUpdate"]>{
-            ...commentUpdateOfVerificationNoSignature,
-            signature: await signCommentUpdateForChallengeVerification({
-                update: commentUpdateOfVerificationNoSignature,
-                signer: this.signer
-            })
-        };
-
-        const toEncrypt = <DecryptedChallengeVerification>{ comment: commentAfterAddingToIpfs.comment, commentUpdate };
-
-        const encrypted = await encryptEd25519AesGcmPublicKeyBuffer(
-            deterministicStringify(toEncrypt),
-            this.signer.privateKey,
-            request.signature.publicKey
-        );
-
-        return { ...toEncrypt, encrypted };
-    }
-
-    private async _publishChallengeVerification(
-        challengeResult: Pick<ChallengeVerificationMessageType, "challengeErrors" | "challengeSuccess" | "reason">,
-        request: DecryptedChallengeRequestMessageType,
-        pendingApproval?: boolean
-    ) {
-        const log = Logger("pkc-js:local-community:_publishChallengeVerification");
-        if (!challengeResult.challengeSuccess) return this._publishFailedChallengeVerification(challengeResult, request.challengeRequestId);
-        else {
-            // Challenge has passed, we store the publication (except if there's an issue with the publication)
-            // call below could fail if the comment is duplicated
-            let failureReason: string | undefined;
-            let toEncrypt:
-                | (DecryptedChallengeVerification & Required<Pick<DecryptedChallengeVerificationMessageType, "encrypted">>)
-                | undefined;
-
-            try {
-                toEncrypt = await this._storePublicationAndEncryptForChallengeVerification(request, pendingApproval);
-            } catch (e) {
-                failureReason = (e as PKCError).message;
-                log.error("Failed to store store Publication And Encrypt For ChallengeVerification", e);
-            }
-
-            const toSignMsg: Omit<ChallengeVerificationMessageType, "signature"> = cleanUpBeforePublishing({
-                type: "CHALLENGEVERIFICATION",
-                challengeRequestId: request.challengeRequestId,
-                encrypted: toEncrypt?.encrypted, // could be undefined
-                challengeErrors: challengeResult.challengeErrors,
-                userAgent: this._pkc.userAgent,
-                protocolVersion: env.PROTOCOL_VERSION,
-                timestamp: timestamp(),
-                ...(failureReason ? { reason: failureReason, challengeSuccess: false } : { challengeSuccess: true, reason: undefined })
-            });
-            const challengeVerification = <ChallengeVerificationMessageType>{
-                ...toSignMsg,
-                signature: await signChallengeVerification({ challengeVerification: toSignMsg, signer: this.signer })
-            };
-
-            const pubsubClient = this._clientsManager.getDefaultKuboPubsubClient();
-
-            this._clientsManager.updateKuboRpcPubsubState("publishing-challenge-verification", pubsubClient.url);
-
-            if (!this._challengeExchangesFromLocalPublishers[request.challengeRequestId.toString()])
-                await this._clientsManager.pubsubPublish(pubsubTopicWithfallback(this), challengeVerification);
-
-            this._clientsManager.updateKuboRpcPubsubState("waiting-challenge-requests", pubsubClient.url);
-
-            const objectToEmit = <DecryptedChallengeVerificationMessageType>{ ...challengeVerification, ...toEncrypt };
-            this.emit("challengeverification", objectToEmit);
-            this._ongoingChallengeExchanges.delete(request.challengeRequestId.toString());
-            delete this._challengeExchangesFromLocalPublishers[request.challengeRequestId.toString()];
-            this._cleanUpChallengeAnswerPromise(request.challengeRequestId.toString());
-            log.trace(
-                `Published ${challengeVerification.type} over pubsub topic ${pubsubTopicWithfallback(this)}:`,
-                remeda.omit(objectToEmit, ["signature", "encrypted", "challengeRequestId"])
-            );
-        }
-    }
-
-    private async _parseChallengeRequestPublicationOrRespondWithFailure(
-        request: ChallengeRequestMessageType,
-        decryptedRawString: string
-    ): Promise<DecryptedChallengeRequest> {
-        let decryptedJson: DecryptedChallengeRequest;
-        try {
-            decryptedJson = parseJsonWithPKCErrorIfFails(decryptedRawString);
-        } catch (e) {
-            await this._publishFailedChallengeVerification(
-                { reason: messages.ERR_REQUEST_ENCRYPTED_IS_INVALID_JSON_AFTER_DECRYPTION },
-                request.challengeRequestId
-            );
-            throw e;
-        }
-
-        const parseRes = DecryptedChallengeRequestSchema.loose().safeParse(decryptedJson);
-        if (!parseRes.success) {
-            await this._publishFailedChallengeVerification(
-                { reason: messages.ERR_REQUEST_ENCRYPTED_HAS_INVALID_SCHEMA_AFTER_DECRYPTING },
-                request.challengeRequestId
-            );
-
-            throw new PKCError("ERR_REQUEST_ENCRYPTED_HAS_INVALID_SCHEMA_AFTER_DECRYPTING", {
-                decryptedJson,
-                schemaError: parseRes.error
-            });
-        }
-
-        return decryptedJson;
-    }
-
-    private _buildRuntimeChallengeRequestPublication({
-        publication,
-        authorCommunity
-    }: {
-        publication: PublicationFromDecryptedChallengeRequest;
-        authorCommunity?: PublicationWithCommunityAuthorFromDecryptedChallengeRequest["author"]["community"];
-    }): PublicationWithCommunityAuthorFromDecryptedChallengeRequest {
-        return {
-            ...publication,
-            author: buildRuntimeAuthor({
-                author: publication.author,
-                signaturePublicKey: publication.signature.publicKey,
-                community: authorCommunity
-            })
-        };
-    }
-
-    private _buildRuntimeChallengeRequest({
-        request,
-        authorCommunity
-    }: {
-        request: DecryptedChallengeRequestMessageType;
-        authorCommunity?: PublicationWithCommunityAuthorFromDecryptedChallengeRequest["author"]["community"];
-    }): DecryptedChallengeRequestMessageTypeWithCommunityAuthor {
-        // This function needs to be updated everytime we add a new publication type
-        const runtimeRequest = remeda.clone(request) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor;
-
-        if (request.comment)
-            runtimeRequest.comment = this._buildRuntimeChallengeRequestPublication({
-                publication: request.comment,
-                authorCommunity
-            }) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor["comment"];
-        if (request.vote)
-            runtimeRequest.vote = this._buildRuntimeChallengeRequestPublication({
-                publication: request.vote,
-                authorCommunity
-            }) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor["vote"];
-        if (request.commentEdit)
-            runtimeRequest.commentEdit = this._buildRuntimeChallengeRequestPublication({
-                publication: request.commentEdit,
-                authorCommunity
-            }) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor["commentEdit"];
-        if (request.commentModeration)
-            runtimeRequest.commentModeration = this._buildRuntimeChallengeRequestPublication({
-                publication: request.commentModeration,
-                authorCommunity
-            }) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor["commentModeration"];
-        if (request.communityEdit)
-            runtimeRequest.communityEdit = this._buildRuntimeChallengeRequestPublication({
-                publication: request.communityEdit,
-                authorCommunity
-            }) as DecryptedChallengeRequestMessageTypeWithCommunityAuthor["communityEdit"];
-
-        return runtimeRequest;
-    }
-
     async handleChallengeRequest(request: ChallengeRequestMessageType, isLocalPublisher: boolean) {
-        const log = Logger("pkc-js:local-community:handleChallengeRequest");
-
-        if (this._ongoingChallengeExchanges.has(request.challengeRequestId.toString())) {
-            log("Received a duplicate challenge request", request.challengeRequestId.toString());
-            return; // This is a duplicate challenge request
-        }
-        if (isLocalPublisher) {
-            // we need to mark the challenge exchange as ongoing for local publishers and skip publishing it over pubsub
-            log("Marking challenge exchange as ongoing for local publisher");
-            this._challengeExchangesFromLocalPublishers[request.challengeRequestId.toString()] = true;
-        }
-        this._ongoingChallengeExchanges.set(request.challengeRequestId.toString(), true);
-        const requestSignatureValidation = await verifyChallengeRequest({ request, validateTimestampRange: true });
-        if (!requestSignatureValidation.valid)
-            throw new PKCError(getErrorCodeFromMessage(requestSignatureValidation.reason), {
-                challengeRequest: remeda.omit(request, ["encrypted"])
-            });
-
-        const decryptedRawString = await this._decryptOrRespondWithFailure(request);
-
-        const decryptedRequest = await this._parseChallengeRequestPublicationOrRespondWithFailure(request, decryptedRawString);
-
-        const publicationFieldNames = remeda.keys.strict(DecryptedChallengeRequestPublicationSchema.shape);
-        let publication: PublicationFromDecryptedChallengeRequest;
-        try {
-            publication = derivePublicationFromChallengeRequest(decryptedRequest);
-        } catch {
-            return this._publishFailedChallengeVerification(
-                { reason: messages.ERR_CHALLENGE_REQUEST_ENCRYPTED_HAS_NO_PUBLICATION_AFTER_DECRYPTING },
-                request.challengeRequestId
-            );
-        }
-        let publicationCount = 0;
-        publicationFieldNames.forEach((pubField) => {
-            if (pubField in decryptedRequest) publicationCount++;
-        });
-        if (publicationCount > 1)
-            return this._publishFailedChallengeVerification(
-                { reason: messages.ERR_CHALLENGE_REQUEST_ENCRYPTED_HAS_MULTIPLE_PUBLICATIONS_AFTER_DECRYPTING },
-                request.challengeRequestId
-            );
-
-        // Reject deprecated wire format fields early, before signature verification
-        // (these fields are never in signedPropertyNames and would otherwise fail with a generic error)
-        if ("subplebbitAddress" in publication) {
-            return this._publishFailedChallengeVerification(
-                { reason: messages.ERR_PUBLICATION_USES_DEPRECATED_SUBPLEBBIT_ADDRESS },
-                request.challengeRequestId
-            );
-        }
-        if ("communityAddress" in publication) {
-            return this._publishFailedChallengeVerification(
-                { reason: messages.ERR_PUBLICATION_USES_DEPRECATED_COMMUNITY_ADDRESS },
-                request.challengeRequestId
-            );
-        }
-
-        const authorSignerAddress = await getPKCAddressFromPublicKey(publication.signature.publicKey);
-        const authorDomain = getAuthorNameFromWire(publication.author);
-
-        // Check publication props validity
-        const communityAuthor = this._dbHandler.queryCommunityAuthor(authorSignerAddress, authorDomain);
-        const decryptedRequestMsg = <DecryptedChallengeRequestMessageType>{ ...request, ...decryptedRequest };
-        const decryptedRequestWithCommunityAuthor = this._buildRuntimeChallengeRequest({
-            request: decryptedRequestMsg,
-            authorCommunity: communityAuthor
-        });
-
-        try {
-            await respondWithErrorIfSignatureOfPublicationIsInvalid(this, decryptedRequestMsg); // This function will throw an error if signature is invalid
-        } catch (e) {
-            log.error(
-                "Signature of challengerequest.publication is invalid, emitting an error event and aborting the challenge exchange",
-                e
-            );
-            this.emit("challengerequest", decryptedRequestWithCommunityAuthor);
-            return;
-        }
-
-        log.trace("Received a valid challenge request", decryptedRequestWithCommunityAuthor);
-
-        this.emit("challengerequest", decryptedRequestWithCommunityAuthor);
-
-        const publicationInvalidityReason = await checkPublicationValidity(this, decryptedRequestMsg, publication, communityAuthor);
-        if (publicationInvalidityReason) {
-            if (DUPLICATE_PUBLICATION_ERRORS.has(publicationInvalidityReason)) {
-                const sig = publication.signature.signature;
-                const attempts = (this._duplicatePublicationAttempts.get(sig) || 0) + 1;
-                this._duplicatePublicationAttempts.set(sig, attempts);
-                if (attempts <= 1) {
-                    return this._publishIdempotentDuplicateVerification(
-                        decryptedRequestMsg,
-                        request.challengeRequestId,
-                        publicationInvalidityReason
-                    );
-                }
-            }
-            return this._publishFailedChallengeVerification({ reason: publicationInvalidityReason }, request.challengeRequestId);
-        }
-
-        const answerPromiseKey = decryptedRequestWithCommunityAuthor.challengeRequestId.toString();
-        const getChallengeAnswers: GetChallengeAnswers = async (challenges) => {
-            // ...get challenge answers from user. e.g.:
-            // step 1. community publishes challenge pubsub message with `challenges` provided in argument of `getChallengeAnswers`
-            // step 2. community waits for challenge answer pubsub message with `challengeAnswers` and then returns `challengeAnswers`
-            await this._publishChallenges(challenges, decryptedRequestWithCommunityAuthor);
-            const challengeAnswerPromise = new Promise<DecryptedChallengeAnswer["challengeAnswers"]>((resolve, reject) =>
-                this._challengeAnswerResolveReject.set(answerPromiseKey, { resolve, reject })
-            );
-            this._challengeAnswerPromises.set(answerPromiseKey, challengeAnswerPromise);
-            const challengeAnswers = await this._challengeAnswerPromises.get(answerPromiseKey);
-            if (!challengeAnswers) throw Error("Failed to retrieve challenge answers from promise. This is a critical error");
-            this._cleanUpChallengeAnswerPromise(answerPromiseKey);
-            return challengeAnswers;
-        };
-        // NOTE: we try to get challenge verification immediately after receiving challenge request
-        // because some challenges are automatic and skip the challenge message
-        let challengeVerification: Awaited<ReturnType<typeof getChallengeVerification>> & { reason?: string };
-        try {
-            challengeVerification = await getChallengeVerification({
-                challengeRequestMessage: decryptedRequestWithCommunityAuthor,
-                community: this,
-                getChallengeAnswers
-            });
-        } catch (e) {
-            // getChallengeVerification will throw if one of the getChallenge function throws, which indicates a bug with the challenge script
-            // notify the community owner that that one of his challenge is misconfigured via an error event
-            log.error("getChallenge failed, the community owner needs to check the challenge code. The error is: ", e);
-            this.emit("error", <PKCError>e);
-
-            // notify the author that his publication wasn't published because the community is misconfigured
-            challengeVerification = {
-                challengeSuccess: false,
-                reason: `One of the community challenges is misconfigured: ${(<Error>e).message}`
-            };
-        }
-
-        await this._publishChallengeVerification(challengeVerification, decryptedRequestMsg, challengeVerification.pendingApproval);
-    }
-
-    private _cleanUpChallengeAnswerPromise(challengeRequestIdString: string) {
-        this._challengeAnswerPromises.delete(challengeRequestIdString);
-        this._challengeAnswerResolveReject.delete(challengeRequestIdString);
-        delete this._challengeExchangesFromLocalPublishers[challengeRequestIdString];
-    }
-
-    private async _parseChallengeAnswerOrRespondWithFailure(challengeAnswer: ChallengeAnswerMessageType, decryptedRawString: string) {
-        let parsedJson: any;
-
-        try {
-            parsedJson = parseJsonWithPKCErrorIfFails(decryptedRawString);
-        } catch (e) {
-            await this._publishFailedChallengeVerification(
-                { reason: messages.ERR_CHALLENGE_ANSWER_IS_INVALID_JSON },
-                challengeAnswer.challengeRequestId
-            );
-            throw e;
-        }
-
-        try {
-            return parseDecryptedChallengeAnswerWithPKCErrorIfItFails(parsedJson);
-        } catch (e) {
-            await this._publishFailedChallengeVerification(
-                { reason: messages.ERR_CHALLENGE_ANSWER_IS_INVALID_SCHEMA },
-                challengeAnswer.challengeRequestId
-            );
-            throw e;
-        }
+        return handleChallengeRequestFreeFunction(this, request, isLocalPublisher);
     }
 
     async handleChallengeAnswer(challengeAnswer: ChallengeAnswerMessageType) {
-        const log = Logger("pkc-js:local-community:handleChallengeAnswer");
-
-        if (!this._ongoingChallengeExchanges.has(challengeAnswer.challengeRequestId.toString()))
-            // Respond with error to answers without challenge request
-            return this._publishFailedChallengeVerification(
-                { reason: messages.ERR_CHALLENGE_ANSWER_WITH_NO_CHALLENGE_REQUEST },
-                challengeAnswer.challengeRequestId
-            );
-        const answerSignatureValidation = await verifyChallengeAnswer({ answer: challengeAnswer, validateTimestampRange: true });
-
-        if (!answerSignatureValidation.valid) {
-            this._cleanUpChallengeAnswerPromise(challengeAnswer.challengeRequestId.toString());
-            this._ongoingChallengeExchanges.delete(challengeAnswer.challengeRequestId.toString());
-            delete this._challengeExchangesFromLocalPublishers[challengeAnswer.challengeRequestId.toString()];
-            throw new PKCError(getErrorCodeFromMessage(answerSignatureValidation.reason), { challengeAnswer });
-        }
-
-        const decryptedRawString = await this._decryptOrRespondWithFailure(challengeAnswer);
-
-        const decryptedAnswers = await this._parseChallengeAnswerOrRespondWithFailure(challengeAnswer, decryptedRawString);
-
-        const decryptedChallengeAnswerPubsubMessage = <DecryptedChallengeAnswerMessageType>{ ...challengeAnswer, ...decryptedAnswers };
-
-        this.emit("challengeanswer", decryptedChallengeAnswerPubsubMessage);
-
-        const challengeAnswerPromise = this._challengeAnswerResolveReject.get(challengeAnswer.challengeRequestId.toString());
-
-        if (!challengeAnswerPromise)
-            throw Error("The challenge answer promise is undefined, there is an issue with challenge. This is a critical error");
-
-        challengeAnswerPromise.resolve(decryptedChallengeAnswerPubsubMessage.challengeAnswers);
+        return handleChallengeAnswerFreeFunction(this, challengeAnswer);
     }
 
     async handleChallengeExchange(pubsubMsg: IpfsHttpClientPubsubMessage) {
-        const log = Logger("pkc-js:local-community:handleChallengeExchange");
-
-        const timeReceived = timestamp();
-
-        const pubsubKilobyteSize = Buffer.byteLength(pubsubMsg.data) / 1000;
-        if (pubsubKilobyteSize > 80) {
-            log.error(`Received a pubsub message at (${timeReceived}) with size of ${pubsubKilobyteSize}. Silently dropping it`);
-            return;
-        }
-
-        let decodedMsg: any;
-
-        try {
-            decodedMsg = cborg.decode(pubsubMsg.data);
-        } catch (e) {
-            log.error(`Failed to decode pubsub message received at (${timeReceived})`, (<Error>e).toString());
-            return;
-        }
-
-        const pubsubSchemas = [
-            ChallengeRequestMessageSchema.loose(),
-            ChallengeMessageSchema.loose(),
-            ChallengeAnswerMessageSchema.loose(),
-            ChallengeVerificationMessageSchema.loose()
-        ];
-
-        let parsedPubsubMsg:
-            | ChallengeRequestMessageType
-            | ChallengeMessageType
-            | ChallengeAnswerMessageType
-            | ChallengeVerificationMessageType
-            | undefined;
-        for (const pubsubSchema of pubsubSchemas) {
-            const parseRes = pubsubSchema.safeParse(decodedMsg);
-            if (parseRes.success) {
-                parsedPubsubMsg = parseRes.data;
-                break;
-            }
-        }
-
-        if (!parsedPubsubMsg) {
-            log.error(`Failed to parse the schema of pubsub message received at (${timeReceived})`, decodedMsg);
-            return;
-        }
-
-        if (parsedPubsubMsg.type === "CHALLENGE" || parsedPubsubMsg.type === "CHALLENGEVERIFICATION") {
-            log.trace(
-                `Received a pubsub message that is not meant to by processed by the community - ${parsedPubsubMsg.type}. Will ignore it`
-            );
-            return;
-        } else if (parsedPubsubMsg.type === "CHALLENGEREQUEST") {
-            try {
-                await this.handleChallengeRequest(parsedPubsubMsg, false);
-            } catch (e) {
-                log.error(`Failed to process challenge request message received at (${timeReceived})`, e);
-                this._dbHandler.rollbackTransaction();
-            }
-        } else if (parsedPubsubMsg.type === "CHALLENGEANSWER") {
-            try {
-                await this.handleChallengeAnswer(parsedPubsubMsg);
-            } catch (e) {
-                log.error(`Failed to process challenge answer message received at (${timeReceived})`, e);
-                this._dbHandler.rollbackTransaction();
-            }
-        }
+        return handleChallengeExchangeFreeFunction(this, pubsubMsg);
     }
 
     async _addCommentRowToIPFS(unpinnedCommentRow: CommentsTableRow, log: Logger) {
